catalog: rename hydrateddesc to hydrateddesccache
This commit only renames things and moves code around. The
hydrateddesccache package now contains only the parts of hydrateddesc
which pertain to the cache data structure, the rest of the hydration
logic has been moved to descs because it is only ever used by the
Collection.

Release note: None
