+++ pkg/BUILD.bazel
@@ -292,7 +292,7 @@ ALL_TESTS = [
     "//pkg/sql/catalog/descpb:descpb_test",
     "//pkg/sql/catalog/descs:descs_test",
     "//pkg/sql/catalog/funcdesc:funcdesc_test",
-    "//pkg/sql/catalog/hydrateddesc:hydrateddesc_test",
+    "//pkg/sql/catalog/hydrateddesccache:hydrateddesccache_test",
     "//pkg/sql/catalog/internal/catkv:catkv_test",
     "//pkg/sql/catalog/internal/validate:validate_test",
     "//pkg/sql/catalog/lease:lease_test",
@@ -1425,8 +1425,8 @@ GO_TARGETS = [
     "//pkg/sql/catalog/fetchpb:fetchpb",
     "//pkg/sql/catalog/funcdesc:funcdesc",
     "//pkg/sql/catalog/funcdesc:funcdesc_test",
-    "//pkg/sql/catalog/hydrateddesc:hydrateddesc",
-    "//pkg/sql/catalog/hydrateddesc:hydrateddesc_test",
+    "//pkg/sql/catalog/hydrateddesccache:hydrateddesccache",
+    "//pkg/sql/catalog/hydrateddesccache:hydrateddesccache_test",
     "//pkg/sql/catalog/ingesting:ingesting",
     "//pkg/sql/catalog/internal/catkv:catkv",
     "//pkg/sql/catalog/internal/catkv:catkv_test",
@@ -2696,7 +2696,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/sql/catalog/desctestutils:get_x_data",
     "//pkg/sql/catalog/fetchpb:get_x_data",
     "//pkg/sql/catalog/funcdesc:get_x_data",
-    "//pkg/sql/catalog/hydrateddesc:get_x_data",
+    "//pkg/sql/catalog/hydrateddesccache:get_x_data",
     "//pkg/sql/catalog/ingesting:get_x_data",
     "//pkg/sql/catalog/internal/catkv:get_x_data",
     "//pkg/sql/catalog/internal/validate:get_x_data",
